nestcam_processsor_v2/requirements.txt
# requirements.txt
streamlit>=1.28.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
moviepy>=1.0.3
pydantic>=2.5.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.0.0
scikit-image>=0.21.0
matplotlib>=3.8.0
plotly>=5.17.0
pandas>=2.1.0
sqlalchemy>=2.0.0
alembic>=1.12.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
mypy>=1.7.0